Moved plotting code to new dir, separated out saving the illusory tuning curves to new file
